requirements.txt
python-dotenv>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0
python-binance>=1.0.17
scipy>=1.7.0
schedule>=1.1.0
pytz>=2021.1
